healthtech/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1

# Database
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9   # PostgreSQL driver

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic[email]==2.9.2

# Optional (for dev convenience)
alembic==1.13.2   # migrations
python-dotenv==1.0.1   # env variables
